break the visible dependency on base
