+++ build.gradle
@@ -4,7 +4,6 @@ buildscript {
     repositories {
         google()
         jcenter()
-        maven { url  "https://dl.bintray.com/kingst/CardScanBase" }
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.4.1'
@@ -24,6 +23,5 @@ allprojects {
     repositories {
         google()
         jcenter()
-        maven { url  "https://dl.bintray.com/kingst/CardScanBase" }
     }
 }
